pyproject.toml
[tool.poetry]
name = "mle-dvc"
version = "0.1.0"
description = ""
authors = ["Kozlov Pavel Alekseevich <kozlovpa@ozon.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.11"
aiobotocore = "2.11.1"
aiohttp = "3.9.3"
aiohttp-retry = "2.8.3"
aioitertools = "0.11.0"
aiosignal = "1.3.1"
amqp = "5.2.0"
annotated-types = "0.6.0"
antlr4-python3-runtime = "4.9.3"
appdirs = "1.4.4"
async-timeout = "4.0.3"
asyncssh = "2.14.2"
atpublic = "4.0"
attrs = "23.2.0"
billiard = "4.2.0"
boto3 = "1.34.27"
botocore = "1.34.27"
catboost = "1.2.2"
category-encoders = "2.6.3"
celery = "5.3.6"
certifi = "2023.11.17"
cffi = "1.16.0"
charset-normalizer = "3.3.2"
click = "8.1.7"
click-didyoumean = "0.3.0"
click-plugins = "1.1.1"
click-repl = "0.3.0"
colorama = "0.4.6"
configobj = "5.0.8"
contourpy = "1.2.0"
cryptography = "42.0.1"
cycler = "0.12.1"
dictdiffer = "0.9.0"
diskcache = "5.6.3"
distro = "1.9.0"
dpath = "2.1.6"
dulwich = "0.21.7"
dvc = "3.42.0"
dvc-data = "3.8.0"
dvc-http = "2.32.0"
dvc-objects = "3.0.6"
dvc-render = "1.0.1"
dvc-s3 = "3.0.1"
dvc-studio-client = "0.18.0"
dvc-task = "0.3.0"
entrypoints = "0.4"
filelock = "3.13.1"
flatten-dict = "0.4.2"
flufl-lock = "7.1.1"
fonttools = "4.47.2"
frozenlist = "1.4.1"
fsspec = "2023.12.2"
funcy = "2.0"
gitdb = "4.0.11"
gitpython = "3.1.41"
grandalf = "0.8"
graphviz = "0.20.1"
greenlet = "3.0.3"
gto = "1.6.2"
hydra-core = "1.3.2"
idna = "3.6"
iterative-telemetry = "0.0.8"
jmespath = "1.0.1"
joblib = "1.3.2"
kiwisolver = "1.4.5"
kombu = "5.3.5"
markdown-it-py = "3.0.0"
matplotlib = "3.8.2"
mdurl = "0.1.2"
multidict = "6.0.4"
networkx = "3.2.1"
numpy = "1.26.3"
omegaconf = "2.3.0"
orjson = "3.9.12"
packaging = "23.2"
pandas = "2.2.0"
pathspec = "0.12.1"
patsy = "0.5.6"
pillow = "10.2.0"
platformdirs = "3.11.0"
plotly = "5.18.0"
prompt-toolkit = "3.0.43"
psutil = "5.9.8"
psycopg2-binary = "2.9.9"
pycparser = "2.21"
pydantic = "2.6.0"
pydantic-core = "2.16.1"
pydot = "2.0.0"
pygit2 = "1.14.0"
pygments = "2.17.2"
pygtrie = "2.5.0"
pyparsing = "3.1.1"
python-dateutil = "2.8.2"
python-dotenv = "1.0.1"
pytz = "2023.4"
pyyaml = "6.0.1"
requests = "2.31.0"
rich = "13.7.0"
ruamel-yaml = "0.18.5"
ruamel-yaml-clib = "0.2.8"
s3fs = "2023.12.2"
s3transfer = "0.10.0"
scikit-learn = "1.4.0"
scipy = "1.12.0"
scmrepo = "2.1.1"
semver = "3.0.2"
shortuuid = "1.0.11"
shtab = "1.6.5"
six = "1.16.0"
smmap = "5.0.1"
sqlalchemy = "2.0.25"
sqltrie = "0.11.0"
statsmodels = "0.14.1"
tabulate = "0.9.0"
tenacity = "8.2.3"
threadpoolctl = "3.2.0"
tomlkit = "0.12.3"
tqdm = "4.66.1"
typer = "0.9.0"
typing-extensions = "4.9.0"
tzdata = "2023.4"
urllib3 = "2.0.7"
vine = "5.1.0"
voluptuous = "0.14.1"
wcwidth = "0.2.13"
wrapt = "1.16.0"
yarl = "1.9.4"
zc-lockfile = "3.0.post1"

[tool.poetry.group.lint]
optional = true

[tool.poetry.group.lint.dependencies]
ruff = "^0.3.0"
mypy = "^1.8.0"


[tool.poetry.group.tests]
optional = true

[tool.poetry.group.tests.dependencies]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120
select = ['N', 'B', 'A', 'E', 'W', 'F', 'ASYNC', 'S', 'D', 'PT', 'I', 'T', 'Q', 'ERA', 'UP', 'COM', 'BLE', 'FA', 'RSE', 'SIM', 'C']
ignore = ['E203', 'PT001', 'D100', 'D101', 'D102', 'D103', 'D104', 'D106', 'D107',  'W293', 'S101', 'D212', 'D200', 'UP007']
extend-exclude = ['*_pb2*.py', '.git', '__pycache__', 'build', 'dist', 'cameras']

[tool.ruff.mccabe]
max-complexity = 10

[tool.ruff.pydocstyle]
convention = 'google'

[tool.ruff.flake8-builtins]
builtins-ignorelist = ["id"]

[tool.mypy]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
addopts = "-s -v"


[tool.black]
line-length = 120
extend-exclude = """
(
  stubs
  | .*pb2.*
  | \\.git
  | __pycache__
  | build
  | dist
  | __init__\\.py
)
"""
